Don't alter kill-ring
Running py-autopep8 will alter the kill-ring leading to some unexpected side effects (for example, if buffers are set to autosave). It makes it confusing to copy a piece of text, move to another buffer and try to paste if py-autopep8 has run in-between. 

This is a bit of a naive quick fix, that simply pops the kill-ring when a kill command is issued. I think a more proper way would be to use `delete-region` to replace `kill-whole-line` and `kill-buffer` usage, but my elisp is rusty and I don't know the equivalents off-hand.

This also affects py-yapf and py-isort, but I won't submit any changes unless you think the current change is sufficient or if something like `delete-region` would be cleaner.